run cc on ConjureGeneratorTaskTest
